Added Timer5 code to make the ChronoDot RTC be the clock used for the sample rate and time/date stamps. It seems to be working though there is a slight drift - it's too slow compared to my computer's clock, so it loses 1 second every 16 or 17 minutes or so. I also repurposed Timer1 to make it control the inSerialSafeRegion, though maybe it would be better to do this in the main loop or other ISRs by checking the value of TCNT5. It isn't fully implemented (still need to disable other interrupts during that period) and is untested. Also, for some reason nothing is getting written to the data file on the SD card anymore.

+++ gcc_lca_mega/fileIO.h
@@ -17,7 +17,7 @@ public:
   String package_name;        // Package name
   unsigned long test_duration;// Test duration
   unsigned int start_delay;   // Selay in seconds between the button push and the test start
-  double sample_rate;         // Sample rate
+  float sample_rate;          // Sample rate
   char temp_units;            // C, F, or K
   String initial_date;        // Initial date
   String initial_time;        // Initial time
